feat: merge generate_task into main.py cli
